threading for rfid polling and camera open
