+++ app/facial_recognition/attendance_controller.py
@@ -1,0 +1,8 @@
+from threading import Event
+
+# Event to control RFID polling and facial recognition
+rfid_event = Event()
+rfid_event.set()  # Start with RFID scanning enabled
+
+# Event to control when to end attendance
+end_event = Event()
